% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegSUIT-package.R
\docType{package}
\name{ggsegSUIT-package}
\alias{ggsegSUIT}
\alias{ggsegSUIT-package}
\title{ggsegSUIT: SUIT Cerebellar Lobular Atlas for the 'ggseg' Ecosystem}
\description{
SUIT cerebellar parcellation into anatomical lobules and deep nuclei for the 'ggseg' ecosystem. Provides a unified 'ggseg_atlas' object with both 2D polygon geometry and 3D vertex indices, for use with 'ggseg' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggseg/ggsegSUIT}
  \item Report bugs at \url{https://github.com/ggseg/ggsegSUIT/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

}
\keyword{internal}
