% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suit}
\alias{suit}
\title{SUIT Cerebellar Lobular Atlas}
\usage{
suit()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Cerebellar parcellation into anatomical lobules and deep nuclei from
the SUIT atlas (Diedrichsen et al., 2009). Contains 2D polygon geometry
and 3D meshes.
}
\examples{
suit()
}
\references{
Diedrichsen J et al. (2009). A probabilistic MR atlas of the
human cerebellum. \doi{10.1016/j.neuroimage.2009.01.045}
}
\concept{ggseg_atlases}
